Feafactor: Move AutoComplete filter function as utils
