+++ pinry-spa/src/components/utils/AutoComplete.js
@@ -1,0 +1,15 @@
+function getFilteredOptions(options, filterText) {
+  return options.filter(
+    (option) => {
+      const index = option.name
+        .toString()
+        .toLowerCase()
+        .indexOf(filterText.toLowerCase());
+      return index >= 0;
+    },
+  );
+}
+
+export default {
+  getFilteredOptions,
+};
